rename mobilewalletlib to dcrlibwallet
