+++ app/walletmediums/dcrlibwallet/walletfunctions.go
@@ -1,4 +1,4 @@
-package mobilewalletlib
+package dcrlibwallet
 
 import (
 	"bytes"
